backend/requirements.txt
asgiref==3.2.3
Django==3.0.3
django-cors-headers==3.2.1
django-cors-middleware==1.5.0
django-filter==2.2.0
django-mssql-backend==2.7.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
PyJWT==1.7.1
pyodbc==4.0.28
pytz==2019.3
sqlparse==0.3.0
